backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Text to video generation service."
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "pydantic[email]==2.11.4",
    "ray[serve]==2.46.0",
    "torch==2.6.0",
    "diffusers==0.35.0",
    "transformers==4.55.2",
    "accelerate==1.10.0",
    "pillow==11.0.0",
    "numpy==2.1.2",
    "opencv-python==4.12.0.88",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "requests==2.28.1",
    "tqdm==4.67.1",
    "pyyaml==6.0.2",
    "sqlalchemy==2.0.43",
    "minio==7.2.16",
    "ruff==0.11.8",
    "pyjwt==2.10.1",
    "psycopg2-binary==2.9.10",
    "peft==0.17.1",
    "nvidia-cufile-cu12==1.14.1.1",
    "bcrypt==4.3.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiohttp-cors==0.8.1",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "attrs==25.3.0",
    "cachetools==5.5.2",
    "certifi==2022.12.7",
    "cffi==2.0.0",
    "charset-normalizer==2.1.1",
    "click==8.1.8",
    "colorful==0.5.6",
    "distlib==0.3.9",
    "filelock==3.13.1",
    "frozenlist==1.6.0",
    "fsspec==2024.6.1",
    "google-api-core==2.24.2",
    "google-auth==2.40.1",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "grpcio==1.71.0",
    "h11==0.16.0",
    "hf-xet==1.1.8",
    "httptools==0.6.4",
    "huggingface-hub==0.34.4",
    "idna==3.4",
    "importlib-metadata==8.7.0",
    "jinja2==3.1.4",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "markupsafe==2.1.5",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.4.3",
    "networkx==3.3",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "opencensus==0.11.4",
    "opencensus-context==0.1.3",
    "packaging==24.1",
    "platformdirs==4.3.8",
    "prometheus-client==0.21.1",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==6.30.2",
    "psutil==7.0.0",
    "py-spy==0.4.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pycryptodome==3.23.0",
    "pydantic-core==2.33.2",
    "python-dotenv==1.1.0",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "six==1.17.0",
    "smart-open==7.1.0",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "sympy==1.13.1",
    "tokenizers==0.21.4",
    "triton==3.2.0",
    "typing-extensions==4.12.2",
    "typing-inspection==0.4.0",
    "urllib3==1.26.13",
    "uvloop==0.21.0",
    "virtualenv==20.31.1",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.20.0",
    "zipp==3.23.0",
    "pytest>=8.3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "ruff>=0.0.252",
]

[tool.ruff]
line-length = 90
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
# Enable flake8-bugbear rules.
select = ["E", "F", "B", "I", "N", "UP", "ANN", "S", "BLE", "FBT", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "PT", "Q"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
